--- src/components/SideNavigator/index.ts
@@ -1,3 +1,0 @@
-import SideNavigator from './SideNavigator'
-
-export default SideNavigator
